# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dharma API
  description: |
    REST API for Hindu mythology data (deities, slokas, temples, etc.).

    **Versioning:** All paths under `/api/...` are mirrored under `/v1/...` (same handlers).
    This spec documents the `/api` prefix; substitute `/v1` for the same routes.

    **Envelope:** Most JSON endpoints return `{ "success": true, "data": ... }` or
    `{ "success": true, "data": ..., "meta": { "page", "limit", "total" } }`.
    Errors return `{ "success": false, "message": "...", "code": "..." }`.
  version: 0.2.0
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/deities:
    get:
      tags: [Deities]
      summary: List deities
      parameters:
        - name: q
          in: query
          description: Search by name, slug, or alias (substring)
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category; use `all` or omit for no filter
          schema:
            type: string
            example: deva
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeityListResponse"
...
  /api/mythical-beings:
    get:
      tags: [MythicalBeings]
      summary: List mythical beings
      parameters:
        - name: kind
          in: query
          description: Filter by kind (e.g. asura, yaksha, naga, mythical)
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MythicalBeingArrayResponse"
...
components:
...
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
    DeityListItem:
      type: object
      description: Summary row for list endpoint
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        category:
          type: string
        alternateNames:
          type: array
          items:
            type: string
        aliases:
          type: array
          items:
            type: string
        affiliation:
          type: string
          nullable: true
        abode:
          type: string
          nullable: true
    DeityListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/DeityListItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    MythicalBeing:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        kind:
          type: string
        description:
          type: string
          nullable: true
        lore:
          nullable: true
    MythicalBeingArrayResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/MythicalBeing"
...